requirements.txt
playwright==1.52.0
tqdm==4.67.1
python-dotenv==1.1.0
